send/receive user hashtags via AP
