+++ src/remote/activitypub/models/tag.ts
@@ -13,3 +13,14 @@ export type ITag = {
 	icon?: IIcon;
 	identifier?: IIdentifier;
 };
+
+export function extractHashtags(tags: ITag[]) {
+	if (!tags) return [];
+
+	const hashtags = tags.filter(tag => tag.type === 'Hashtag' && typeof tag.name == 'string');
+
+	return hashtags.map(tag => {
+		const m = tag.name.match(/^#(.+)/);
+		return m ? m[1] : null;
+	}).filter(x => x != null);
+}
